updated some code written 3years ago almost, needs to branch and refactor entire puzzlefield which contains too mouch logic atm .
Should also refactor some ways that the basic blocks are handeled, since its not pretty at all.

Did change some of the adding of blocks in puzzlefield, removed an old function that was not in use, changed the way blocks receive their type and forced the blockqueue into the field so we know which types of blocks will be created.
Later on use some sort of blockpool to create some uniform distrobution of blocks.